requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
ecdsa==0.17.0
gunicorn==20.1.0
httpie==2.6.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==8.4.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycodestyle==2.8.0
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
